GTP-23: Refactoriza servicio de login para tenerlo en un archivo independiente
